lint
SCCS-vsn: sys/kern/init_main.c 7.6
SCCS-vsn: sys/kern/tty.c 7.12
SCCS-vsn: sys/kern/uipc_socket.c 7.9

+++ usr/src/sys/kern/uipc_socket.c
@@ -671,8 +671,9 @@ sosetopt(so, level, optname, m0)
 
 			case SO_SNDBUF:
 			case SO_RCVBUF:
-				if (sbreserve(optname == SO_SNDBUF ? &so->so_snd :
-				    &so->so_rcv, *mtod(m, int *)) == 0) {
+				if (sbreserve(optname == SO_SNDBUF ?
+				    &so->so_snd : &so->so_rcv,
+				    (u_long) *mtod(m, int *)) == 0) {
 					error = ENOBUFS;
 					goto bad;
 				}
